+++ src/server/modules/mailer/index.js
@@ -1,6 +1,6 @@
-import transporter from './transporter';
+import mailer from './mailer';
 import Feature from '../connector';
 
 export default new Feature({
-  createContextFunc: () => ({ transporter })
+  createContextFunc: () => ({ mailer })
 });
